RestServer: Factor out common server functions/interface into a base class
https://github.com/caikit/caikit/issues/226
