+++ tests/runtime/test_grpc_server.py
@@ -67,7 +67,7 @@ from sample_lib.data_model import (
 from tests.conftest import random_test_id, temp_config
 from tests.fixtures import Fixtures
 from tests.runtime.conftest import runtime_grpc_test_server
-import caikit
+import caikit.interfaces.common
 import sample_lib
 
 ## Helpers #####################################################################
